Added the OOP Drone Controller Notebook
I have simulation solution in the form of a class that wraps around the drone controller and physics object. This uses the previously discussed integration method. I also experimented with an ODE solver. The problem with this is the acceleration and control forces are no longer available. I may be implementing it wrong, if you have any advice that would be appreciated.